Split up test module
